requirements.txt
torch
torchvision
transformers
pillow
faiss-cpu
fastapi
uvicorn
python-multipart